chore: refine, apply, and enforce lints (#431)
* fix: enforce workspace lint inheritance

* wip: fmt v1

* wip: apply autofixable lints

* wip: pass 2

* chore: manual lints

* chore: reward curve lint

* wip: packing lint

* fix: enable await-holding-lock lint

* wip: fix truncation

* fix: fork test timeout

+++ crates/chain/tests/block_production/testing_primitives.rs
@@ -58,7 +58,7 @@ async fn heavy_test_mine_tx() {
         .unwrap();
     irys_node.mine_block().await.unwrap();
     let next_height = irys_node.get_height().await;
-    assert_eq!(next_height, height + 1 as u64);
+    assert_eq!(next_height, height + 1_u64);
     let tx_header = irys_node.get_tx_header(&tx.header.id).unwrap();
     assert_eq!(tx_header, tx.header);
     irys_node.stop().await;
